Beans adjustment are done

+++ src/main/java/com/example/vacationpaymentcalculation/service/VacationService.java
@@ -2,11 +2,12 @@ package com.example.vacationpaymentcalculation.service;
 
 import com.example.vacationpaymentcalculation.models.Vacation;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public interface VacationService {
 
-    Double getVacationPayment(Double salary, LocalDate startDate, LocalDate endDate);
+    BigDecimal getVacationPayment(BigDecimal salary, LocalDate startDate, LocalDate endDate);
 
     int checkWeekEndsDay(Vacation vacation);
 }
